# api-spec.yaml
openapi: 3.0.0
info:
  title: Game Hub Manager API
  description: Detailed API documentation for the Game Hub Manager Interface.
  version: 1.0.0
servers:
  - url: /api
    description: Local API server
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session_token
paths:
...
  /users:
    get:
      tags: [Users]
      summary: List users
      description: Admin/CTO only. Returns list of all users.
      responses:
        200: { description: OK }
    post:
      tags: [Users]
      summary: Create user
      description: Admin/CTO only. Creates a new user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string }
                password: { type: string }
                name: { type: string }
                roles: { type: array, items: { type: string } }
      responses:
        201: { description: Created }
...
  /gcs/files:
    get:
      tags: [GCS]
      summary: List storage files
      parameters:
        - name: path
          in: query
          schema: { type: string }
      responses:
        200: { description: OK }
...